refactor: typing
